requirements.txt
Flask==2.3.3
nltk==3.8.1
textblob==0.17.1
python-docx==0.8.11
PyPDF2==3.0.1
werkzeug==2.3.7
speechrecognition==3.10.0
pyaudio==0.2.11
textstat==0.7.3
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3